+++ priv/repo/migrations/20170822155215_create_comment.exs
@@ -1,0 +1,15 @@
+defmodule Collaboration.Repo.Migrations.CreateComment do
+  use Ecto.Migration
+
+  def change do
+    create table(:comments) do
+      add :text, :string
+      add :idea_id, references(:ideas, on_delete: :nothing)
+      add :user_id, references(:users, on_delete: :nothing)
+      timestamps()
+    end
+
+    create index(:comments, [:idea_id])
+    create index(:comments, [:user_id])
+  end
+end