+++ web/views/admin_view.ex
@@ -1,4 +1,7 @@
 defmodule Collaboration.AdminView do
   use Collaboration.Web, :view
 
+  def navtab_class(conn, action) do
+    if action_name(conn) == action, do: "nav-link active", else: "nav-link"
+  end
 end